django/requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
dotenv==0.9.9
drf-yasg==1.21.10
inflection==0.5.1
Markdown==3.8
packaging==25.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
whitenoise==6.9.0
psycopg2==2.9.10
django-redis>=5.4.0
gunicorn==23.0.0
daphne==4.2.0
pillow>=11.2.1